Enable the revive linter (#26)
This commit makes the code a little stricter through the 'revive'
linter. The work was sponsored by make-server branch, which had a 'TODO'
to ensure all code was 'sufficiently documented'. Rather than try and
figure that out as a human, it is simpler to enable a linter.

However, with the linter enabled, the codebase needs to be reconciled
against it. Given this, this commit also does that.