Stub an implementation of an API
This commit stubs an implementation of the API. It is hardly complete —
rather, it is "just enough" that we can start to build the logic of the
API without worrying as much about the plumbing around it.

== Design Notes
=== Proto

The api is defined via a series of protobuf files. Proto (and, as
mentioned later, gRPC) is binary data specification format. Tooling
around it (e.g. protoc) generates libraries that are designed to parse
and interact with this format.

This makes the API implementations extremely "standardized", allows
users to get up and running quickly via these libraries.

=== Buf

Buf is a toolf or versioning and managing protobuf dependencies, as well
as generating the libraries. It wraps protoc et. al in ways that are
simpler to use than doing it manually.

=== GRPC

In addition to protobuf, a gRPC gateway is exposed. GRPC implements the
Server{} specification of proto, and is well supported by the generation
tooling.

=== OpenAPI (via GRPC Gateway)

While good tooling exists for generating gRPC messages, there are people
who would want to interact with the API with "more standard" tooling,
such as HTTP+JSON.

The GRPC Gateway allows translating messages from HTTP+JSON into a
format understood by gRPC.

=== Interceptors

The application exposes multiple different ways of answering messages,
depending on the protocol that is being requested. This include things
like:

  * HTTP+JSON (via Accept header)
  * GRPC (via Content-Type header, exclusively over H2)
  * H2C (via crafted message)

Each of these things do not work with the "default" handler, and are
instead "intercepted" and "offloaded" to the appropriate handler

=== H2C

One of the interesting challenges of running in CloudRun is that it
doesn't allow TLS to the node (although the traffic is encrypted within
a tunnel), and gRPC (nor go) doesn't play well with HTTP/2 without
encryption.

Go does allow H2C via the "X" libraries, and gRPC works fine with this.
However, this requires specifically to "opt in" to this. Given this, a
new interceptor was written replicating logic that does this in the
"/x/" libraries.

It was reimplemented as otherwise, the code becomes hard to reason about
through the "standard" implementation (wrapping the default handler).
This allows the "interceptor" approach, which is semantically similar
but practically simpler.

No tests were written for this work as I've been working on this for
days, and I just want to ship it. I'll invaraibly come back to this.

=== Tools.go

A New file (tools.go) has been introduced to version specific go tools,
as well as allow they're easy discoverability.