Remove the "redirect" subcommand
Currently, there is a "redirect" subcommand for this application.
However, while doing work to prototype the API, it became clear that it
is possible to have different protocols all exposed on the same port,
and through the same HTTP server. For example,

- The link shortner (default router)
- The HTTP application/json API (via Accept header)
- The gRPC API (via Content-Type header)

Given this, the necessity for multiple servers on different ports is
redundant.

This commit makes the first change designed to make that clear — it
removes the "redirect" subcommand. Future work will continue to modify
the server to (optionally) allow the APIs

== Design Notes
=== What does this mean for api.x40.link

In principle, this means that all use cases can be served from the top
level domain (x40.link). However, its not clear whether this will
*always* be the case. x40.link might (for example) heavily use caching
in future, which would be disasterous for the API use case.

Given this, it is likely that separate routes will be setup for the API,
and disallowed for the root domain *even though* the application can
technically serve it all over the same port.