Replace Gin with the Chi router (#29)
While doing parallel work to implement an API, the Gin framework proved
especially challenging to integrate with other HTTP handlers (e.g. GRPC
or the GRPC Gateway). It was difficult to convert between the HTTP
Handler interface and Gin's "context" approach, and Gin replaced the
way status codes were delivered (breaking compat between the two).

On review, Gin was becoming more of a challenge to work with than the
value it brought. Rather than "opt in" to the entire Gin framework, this
commit replaces Gin with a simpler, closer to the standard library
framework — "Chi". This seems to be well maintained, in fairly broad use
and easier to reason through.

=== Additional Tests

In the initial, gin design there were fewer tests. This commit adds more
test, somehow achieving 100% in the server package.

=== Errors w. Context

One of the things the Gin package did that was nice was centralize error
handling for the API, through the use of "context errors". This commit
replicates a simpler version of the same, allowing requests to end by
setting context on the request.

There is a new, central middleware that will look at all requests
egressing the stack, and if there is an error, write out an appropriate
response. This prevents many painful repetitions of the same content.

=== Problem+JSON

With the use of the Errors library, the opportunity came up to improve
the way erros are displayed. One of the more useful ways — absent a UI —
is to respond with a JSON (or similarly structured) blob.

This commit takes advantage of an existing library that does exactly
this, rather than writing it again.

=== Noroute to Get<slug>

This commit swaps the existing "noroute" design to a design that matches
the first slug in the URL. This is marginally different than the
previous design, which would (by definition) match the entire URL.

This allows, in principle, other bindings later for things like the
.well-known URLs:

  * https://github.com/OAI/OpenAPI-Specification/issues/2128