Add Firestore as a storage backend
Currently the storage that is being used within Google Cloud is YAML
based storage. However, for the application to make any sense, it needs
to be able to dynamically read and write URLs.

This commit implements Google Cloud firestore as a key/value storage
that seems to work appropriately.

== Design Notes
=== Firestore

Choosing firestore was essentially googling "Googel Cloud DynamoDB
equivelant" and then picking what showed up. There seems to be a free
tier, and a default database.

=== Emulator

Testing against firestore means testing against an implementation of it,
which is only (apparently) possible by running an emulated version of
it. This is *slow* — its some JAR Google Cloud runs — so its offloaded
to a new, external compliacne test suite.

=== Changing the default

This commit changes the default for the release in Google Cloud. While
this will probably be annoying for some people, it is quite a motivation
for me to fix it.