Refactor away the handler implementations in interceptor
Currently there is a conceptual problem I'm trying to figure out: How to
limit the handler(s) to respond only on specific "virtual hosts" (or
when the host header matches a request), instead of responding on all.
The XY problem I'm looking to solve is how to:

1. Shift the API queries to go behind an API endpoint (api.x40.link), so
   that x40.link can be safely "heavily cached" (either indefinitely
   with invalidation, or for 5 minutes or so without it).
2. Allow a UI to be embedded at app.x40.link

This would allow x40.link to be sent back to the app, and the API
configured never to be cached.

To do that, we need to extend the Match() function so "matchers can wrap
matchers". So, a virtual host allowlist can happen before other
matchers, thereby allowing us to limit the requests arbitrarily.

== Design Notes
=== H2C

The largest change that happens in this pull request is removing H2C.
This is possible because the only part of H2C we actually need is the
matching part, rather than the rest.