Wait for firestore to come up
Currently there is a problem with firestore coming up — it seems super
slow. This commit seeks to address that by polling the conneciton it is
(supposed) to create, and if it cannot create it, bailing out (via
panic()

This should make the tests less flaky.

== Design Notes
=== Command Output

The command output is persisted, as for some reason its struggling to
run in CI/CD.