Limit API access to specific hosts (#40)
As mentioned in 307eab3, the goal generally is to have the API be behind
api.x40.link. This commit makes that possible by building a new
filtering mechanism to wrap the existing matchers (IsHost) and a new
method to combine matchers (AllOf).

== Design Notes
=== Refactoring Flags

This commit makes an enormous breaking change to the UI — all of the
flags are changed. The reason for this is the new flag design is
simpler, both in terms of the implementation (using the configuration
paths as flags) and in terms of the reasonability (all configuration
flags can be inspected via the CLI).