Remove the gRPC Gateway implementation (#44)
Currently, I'm trying to figure out a nice way to authenticate requests.
This turns out to be challenging with the existing gRPC design —
specifically, the gateway.

The way the gateway is implemented is that it deserializes requests from
JSON into something that the gRPC server implementation expects
(skipping the intermediary layer with proto, as well as a bunch of
other gRPC stuff). This is a reasonable approach, but one of the things
it skips is gRPC interceptors.

In parallel, authentication (as well as debugging and a series of other
use cases) are implemented in GRPC Interceptors. This means that:

1. Authentication would need to be implemented in *both* the gRPC
   gateway design, and the gRPC design
2. The gRPC Gateway would need to be modified to include the
   interceptors (such as proxying all connections over localhost or to a
   buffered connection pool)

Neither seems like the appropriate implementation right now. In
principle, the gRPC gateway could be modified to run the gRPC
interceptors — but it seems this is itself not trivial, as they skip the
intermediary construction of the request.

Given this, this commit opts to avoid the whole problem by ripping out
the gRPC gateway for the time being. There will likely be an answer
here, but rather than pause to figure it out now, the gateway is being
removed.

== Design Notes
=== Harder non-cloud mode

This commit makes the proposed API substantially harder to work in
non-cloud mode, or for developers to poke with tools like cURL. This is
*already* hard for the existing deployment which assumes clients will
handle (in future) OIDC, but for the offline case (e.g. localhost) is
now harder.

Future work will allay these first by building a client that is
especially designed for this, and in the intermediary, people who are
playing can use tools like grpcurl.

=== Potential future avenue

Google Cloud offers something like this via the Cloud Endpoints product:

* https://cloud.google.com/endpoints/docs/grpc/transcoding

But this _essentially_ the same design as proxying it over localhost,
with the exception it "feels better" as "google takes care of" one side
of the connection.