Add missing backend storage tests (#38)
* Add missing backend storage tests

In 4f49d07, work was done to add a firestore backend. However,
inadvertently, the commit missed the created tests.

This commit addresses that.

* Wait for firestore to come up

Currently there is a problem with firestore coming up — it seems super
slow. This commit seeks to address that by polling the conneciton it is
(supposed) to create, and if it cannot create it, bailing out (via
panic()

This should make the tests less flaky.

== Design Notes
=== Command Output

The command output is persisted, as for some reason its struggling to
run in CI/CD.

* Install the firestore emulator in GitHub Actions

Currently, the emulator is missing. This explains why the tests are
failing.

This commit installs it.