Implement authentication (#47)
Currently, there is no authentication for this application. This makes
exposing these APIs dangerous, as whomever can come along, crawl and
fuzz them and (even if inadvertently) end up costing me an enormous
amount of money.

This commit implements a very limited authentication.

== Design Notes
=== Style

This commit implements authentication via the OpenID specification. This
specification is an open standard that others can, in principle, also
reuse with this application if they choose to self-host it somewhere.

See:

1. https://openid.net/specs/openid-connect-core-1_0.html

Google Cloud supports OpenID natively via its Cloud resources:

  * https://developers.google.com/identity/protocols/oauth2

Notably, this is also the authentication used for accessing *google
cloud resources*, which means (in particular) if my token gets exposed,
someone gains the same rights I do to the account. Given this, a new
account with no APIs enabled _except_ oAuth has been created, and the
IDs associated given to that.

Long term, something like Auth0 is likely to be a better fit.

=== No authz

This commit does not do any implementation of authz. Currently, I am not
sure of the best way to do this. Initially, I imagined I'd just validate
against a series of scopes per method — but this does not look trivial
to implement.

Given this, we just use authn. There are also no tests, for the minute.

=== No Ownership

This commit does not solve ownership of records — who has access to one,
has access to all right now. This is not a problem while I am the only
one who has access to the application, but will be when more people do.

This requires a rewrite of the storage layer; likely with an interface
that allows asserting ownership of a given record.