Wire the notion of ownership through the storage implementations
Currently, work is going on to implement an API for this service,
including ensuring that a (multi-tenant) userbase cannot overwrite each
others records. This commit implements the path for a first, naive
approach at this: Introducing ownership to storage engines.

It does this by providing an interface an engine can (optionally)
implement which allows us to check whether a given agent (supplied by
string) is the "owner" of a given record. This, in turn, allows us to
check whether a record should be updatable (or deletable).

== Design Notes
=== Simple Design

This design is deliberately extremely simple. The appropriate solution
looks to be something like ReBAC¹, but this will take a long time to
write manually, and there's nothing yet that balances the desire to have
this open source and widely usable with the desire to host it in Google
Cloud.

Interestingly, Google Cloud itself does not offer a Zanzibar like
approach, despite apparently releasing the idea²

—
1. https://en.wikipedia.org/wiki/Relationship-based_access_control
2. https://research.google/pubs/zanzibar-googles-consistent-global-authorization-system/