Enable Firestore to reject non-ownership requests
Currently, there are no stores that implement the "authenticator"
interface. This commit extends firestore so that it can validate the
ownership of records.

== Design Notes
=== Firebase Enforcement

One of the interesting questions is whether or not the Firebase storage
layer _itself_ should do the enforcement of ACLs. I'm ... not sure yet —
I'm tending toward no as I write this commit message. The alternative is
ensuring that wherever the Firebase storage layer is accessed, its
always up to date.

This implementation of auth is funky. But at least it will help us get
started.