Propagate context to storage
Currently there is a problem in that the context that is created in the
oidc layer is not propagated all the way to the storage layer (and thus,
the storage layer cannot make any decisions about how it is used).

This commit addresses that by wiring through the context some places it
had been missed.