Implement the server (#52)
With the implementation of authentication (and the limit of that
authentication to the andrewhowden.com organization), this project can
(relatively safely) implement the API.

This commit does that.

== Design Notes
=== Renaming Methods

This commit renames the existing methods to be something more "RPC
Friendly". Historicaly, the names were derived from the RESTful HTTP
layer they were trying to emulate.

While thinking about transactions in a RESTful way is still a useful
idea for state management, naming the methods after the HTTP verbs
doesn't make as much sense.

The names are now:

* Get → Get (Still Valid)
* Post → New (The most common method)
* Post → NewCustom (The exception, where we're creating new URLs)

=== Updates to Taskfile

While generating the new definitions, I noticed that the existing design
doesn't let us generate outside a build (e.g. during development), and
during the build, its invoked every time bin is called — it only needs
to be done once.

This commit splits out the task so it can be invoked seperately, and
adds it as a dependency to bin so its only called once.

=== Partial Implementation

This commit only partially implements the desired API — specifically,
the method "POST" is currently not implemented.

This method relies on something generating UUIDs for these URLs, which
hasn't yet been generated.