Implement the "New" function (#53)
This commit implements the new function, within the URL. While doing
this, quite a bit became clear about the design of the API that would
allow it to be simplified.

== Design Notes
=== Removing custom, replacing with enricher

The difference between the New and the NewCustom methods was the
existance of a path. However, what became clear when the New method was
being created was that *both* the path and the slug would need to be
generated. With that both being there, it became clear that there could
be a single function which simply generated the content that was
missing.

The old function was renamed, and the protobuf definitions updated. A
new type (Enricher) was created to polyfill the information that is
missing.

=== Hardcoded x40.link default

The enricher configuration currently has a hard-coded link of x40.link
as a default.

This is just laziness. I need to figure out a way to reduce the
complexity in cmd.Serve(), which I think will be dependency injection.
Then, we can figure out how to dependency inject configuration (perhaps
by storing specific types in Viper).

At that point, the default will be configurable.

=== General design for UID

This commit also implements a simple generator for the slugs. The focus
for this generator is not necessarily uniqueness, but rather forward
compatibility. This is because it is very unlikely that I will figure
out the right algorithm in the initial pass; instead, I want to be able
to evolve the algorithm going forward.

To this end, the algorithm just samples random data, and converts it to
base64 to have a more compact representation of it. Its prefixed with a
type, so that if needed in future, the algorithm can be swapped out and
the old URLs preserved.

=== Testing implements in uid

There are generators in the uid generation, including testing
generators. This just allows easier validation of the UI.