Move authentication to Auth0 (#54)
Currently, the authentication is managed via the Google Cloud identity
provider. However, this has some substantial shortfalls — for example,
it is bound to access the resources within Google Cloud (rather than
resources within an application).

This commit moves the authentication instead to auth0. Auth0 is a
provider that specializes in this approach, implements OIDC as expected
and allows a different style of grant that doesn't need to persist
secret credentials.

== Design Notes
=== Role Validation

This commit implements more (untested) role validation. This is designed
to allow a larger number of users to access the service, but to limit
those users to an allowlist that I manually control.

=== Terraform Auth0

The Auth0 definitions are all included in terraform for management, as
well as transparency for potential employers.