Test the OIDC implementation for the server
Currently, the OIDC implementation for the server does not have unit
tests at all. This is because I didn't really know how OpenID worked —
what exactly it was verifying, and why (despite knowing prety well how
asymmetric cryptography works).

This commit addresses that as I spent the time to figure out exactly
what was being verified, and how to reproduce it.

== Broader Context

The plan, currently, is to extend the auth package in API to also
function as an interceptor for the client implementation of gRPC, rather
than just the server implementation. This might meaning extending its
responsibilities some, but we can do that one step at a time.

== Design Notes
=== Rename VerifyCtx

This commit renames VerifyCtx as there will be (in future) another
function called something like "EnrichCtx", which will include a valid
token in client requests.

=== Stripping Token

While writing unit tests, it occurred to me that the token should
probably be stripped out of the context, so it doesn't leak into other
middleware (e.g. observability middleware).

=== Tokens package

While writing tests, I was learning how tokens worked, their common
claims and so on. This knowledge is preserved via the tokens package,
which lays out the different tokens this service is built on, and
describes them in some depth.

This is not different than the OIDC Tokens, with the exception of the
X40 token. But it is nice to have it documented and clear.

=== General appraoch to testing

The general approach to testing is to mock the verification aspect of
OIDC. Other alternatives I considered were running a mock OIDC server,
but this is pretty high overhead for what is (essentially) a "signature
verification" process, and the package allows us to do that in a much
lighter way via NewVerifier