Implement token generation (via "Seed")
Currently, the broad work stream is to create authentication that can be
used between the client and the server. Part of that work is generating
tokens that the client can send to the server, and then authorize them.

Without going into detail, the plan for this work is to implement a
"TokenSource", and have that "TokenSource" be cache aware. This means it
reads its initial state from disk, and if the disk is empty, needs the
cache to be "seeded".

This commit implements the seed.

== Design Notes
=== No OpenID

This commit is notable as while it generates tokens, it does not
generate tokens that are OpenID compatible. This is because apparently,
OpenID tokens are designed to be used on the client, and authentication
tokens used on the server¹

Auth0 provides support for JWT access tokens, and the existing
enrichment provides the roles.

—
1. https://auth0.com/docs/secure/tokens/access-tokens
2. https://www.oauth.com/oauth2-servers/openid-connect/id-tokens/

=== Auth Flow: Device

The chosen authenticatication flow is the one for non-interactive
devices; namely, the device flow. This is because this flow does not
require a secret key to be persisted in the application; rather, it
generates a "token" that is single use and for the device session. The
user then authenticates on another device.

=== HTTP Testing

This commit also goes some way to figuring out how to test OAuth by
creating a HTTP Transport RoundTripper that simply uses a chi
multiplexer to handle requests. This is very efficient, though the
requests still take a second as the ticker can only be reduced down to a
second.