Implement the auth0 side of device authentication
This commit implements the auth0 side of device authentication.
Specifically, it:

== Creates a new resource ("API")

This commit creates a new resource that can be requested for when
requesting an access token via the "audience" field:

* https://auth0.com/docs/secure/tokens/access-tokens/get-access-tokens#control-access-token-audience

This allows querying, at the token generation time, whether or not a
user has the ability to access scopes defined on that resource.

=== Removing the enrichment of roles

This commit removes the enrichment of the token with the roles assigned
to the user. This was a bastardization of OAuth, and is better served by
using the scopes and resources implementation.

=== Changing grant type

This commit changes the grant type from "authorization" code to the
"Device" style authorization. This is a superior form, especially when
working with the CLI in which the browser (may) not be available.

It additionally has no private keys.

=== Assign scopes to role

This commit assigns scopes that are stored in the application to the
role that people are expected to have if they interact with the site.
This will break all existing auth (which is fine), but will allow a
simpler auth implementation in future.