Replace the OIDC implementation with simple JWT validation (#58)
In parallel work (3d1fb9a), the approach for authentication was changed
from OIDC to authentication tokens from auth0. This commit implements
the server side of this change.

The changes are substantial, and are described in more detail in design
notes.

== Design Notes
=== Permissions

The permissions are now crafted after the method names. This could get
... complex, if there are a lot of methods. However, this seems unlikely
given the limited nature of this application.

Future work should replace this whole design with ReBAC regardless, and
rip the RBAC style out.

There are also "anonymous methods", in which we expect a token but no
role. These are just metadata routes (e.g. reflection API)

=== auth.proto

As part of this work, rather than the permissions being defined in
the file, they're defined in the protobuf definition as a method option.

This is a ... contentious choice. There is substantial hesitation in the
gRPC / Proto repositories for supporting this design, with being
supposably "a property of the application". This might be due to
Google's use of ReBAC as the design of choice (see Zitadel), and thus
not really understanding the RBAC use case.

For now, it was a good excuse to learn how to access options in-process.
In future, it would be good to learn how to have code generated from
these, rather than reading the embedded proto definitions.

=== OIDC to JWT Tokens

The interesting question is: could OIDC have been useful nonetheless? It
certainly solves problems around the expected claims, and rewriting the
ownership of records is likely to be challenging.

However, the testability of the code substantially improves now that the
OIDC dependencies are removed, and the JWT validation is done
"manually". I know far better understand the validation process itself,
and can understand when things are "valid" versus "not".