Dependency inject configuration (#60)
Currently, the entrypoint for the applications (the "cmd.Serve
execution function") is getting enormous. It includes a wide range of
logic, mostly around bootstrapping dependencies.

This commit refactors that out, and uses a dependency injection
framework to supply the required values. This allows distributing the
bootstrap next to the various packages.

== Design Notes
=== Chosen Framework

The chosen dependency injection tooling is "Wire". Wire is unique in
that it _generates the depedency injection_ — This allows it to be used
with a remarkably small amount of tooling; just the initial code
generators.

=== Rewriting Configuration

This commit also substantially rewrites the way configuration is used,
so that the configuration can take care of tasks like binding to a flag
set or storing information about itself. This, in principle, would allow
much simpler generated templates for Yaml and similar work. In practice,
it just makes the cmd entrypoints much simpler, and allows reuse across
commands (e.g. the upcoming client tooling)

=== DI Packages

In a couple of packages, the "root" package provides a series of
interfaces and similar work that the "child" packages then consume. This
means that the "root" page cannot do dependency injection, including
reconciling which "sub" package to use.

Given this, the dependency injection work is occasionally in its own
"sub" package, called "di".

=== Joining with Viper

This commit uses the "...<OptionFunc>" syntax as the signature to supply
with Wire. This allows creating functions like OptionFromViper, which
convert the Viper options into package specific New() function options.

This is a relatively pleasant way to encapsulate this logic.

=== No Tests

This commit doesn't add too many tests for the configuration work,
though in principle that is now possible. While it could be done, this
is still a personal project and I want to get this more toward usable
before being super finicky about coverage.