Introduce k/v storage layer
Currently work is in progress to define a client for the x40
application. This client needs to store an access and a refresh token
_somewhere_, and it cannot store it on memory. this commit provides some
of the content required to keep that storage on disk.

== Design Notes
=== Storage Interface

This commit provides storage via an _interface_, rather than directly
using the file. It turns out, this is a tricky problem in open source
software generally — the goal here is to create a pluggable interface,
and perhaps (in future) another design where we can persist it via
boltdb or similar.