fix(certificates): Allow certificates to automatically renew
Right now, the certificate is serving a domain h4n.link, which expired
(and was not renewed). This blocked renewal.

This commit addresses that by removing this certificate from the
configuration, which should allow the certificate to recover. This will
still be unavailable for a while, as load balancers in Google tend to be
flaky.